feat: support AWS SSM tailscaled state (#41)
## what
* This adds support to store `tailscaled` state in AWS SSM. This helps
reusing device state names rather than deleting old devices which is not
supported by Tailscale via Terraform ATM.
* This adds recent Masterpoint's GH + CRabbit configs.
* This sets trunk update to run less often to reduce noise.
* This adds ability to confugure ASG min/max size and desired capacity.
This is important in case of using an external state to avoid the
[`Duplicate node key `
issue](https://tailscale.com/kb/1023/troubleshooting):
> This can occur if you use a backup of one device to create another, or
clone a file system from one device to another. The Tailscale
configuration files are duplicated. The Tailscale files will need to be
removed from one of the two.
You can identify duplicated devices in the
[Machines](https://login.tailscale.com/admin/machines) page of the admin
console by looking for a Duplicate node key badge underneath the device
name.
On one of the systems, [uninstall and completely
delete](https://tailscale.com/kb/1069/uninstall) the Tailscale app. It
is especially important to remove the files listed for your platform,
the goal is to make a new Tailscale IP address when it is installed
again. Then, [reinstall the app](https://tailscale.com/kb/install).


## why
* Ephemeral nodes behave glitchy during the rotation - we had to perform
a manual instance restart to run Tailscale. It's hard to reproduce, so
keeping a state in external storage in one more option to try to keep
the device in order.

## references
* https://tailscale.com/kb/1278/tailscaled#flags-to-tailscaled



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
- Introduced a new configuration file for CodeRabbit integration,
enhancing review and feedback processes.
- Added new modules for managing SSM state parameters and IAM policies
in the Terraform setup.
- Expanded configuration options with new variables for Auto Scaling
Group and Tailscale state management.
- Updated the workflow for trunk upgrades to run monthly, improving
efficiency.

- **Documentation**
- Enhanced `README.md` with new module and variable details for better
user guidance.
  
- **Chores**
	- Updated `.gitignore` to manage ignored files more effectively.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Matt Gowie <matt@masterpoint.io>